requirements.txt
python-dotenv==1.0.0
streamlit
tiktoken==0.4.0
faiss-cpu==1.7.4
pinecone-client
langchain==0.1.2
InstructorEmbedding
sentence-transformers==2.2.2
langchain-pinecone
langchain-google-genai
streamlit-option-menu
pyperclip
PyPDF2